Worker/requirements.txt
# Function dependencies, for example:
# package>=version
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-pubsub==2.16.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6